fix(vault): remove standalone runtime MCP layer
