+++ .github/workflows/client-ci.yml
@@ -6,7 +6,6 @@ on:
       - "apps/web/**"
       - "packages/shared/**"
       - "packages/cli/**"
-      - "packages/runtime-mcp/**"
       - "packages/whatsapp-baileys-sidecar/**"
       - "backend/app/routes/channel_routers/whatsapp.py"
       - "backend/app/services/whatsapp_*.py"
@@ -32,7 +31,6 @@ on:
       - "apps/web/**"
       - "packages/shared/**"
       - "packages/cli/**"
-      - "packages/runtime-mcp/**"
       - "packages/whatsapp-baileys-sidecar/**"
       - "backend/app/routes/channel_routers/whatsapp.py"
       - "backend/app/services/whatsapp_*.py"
@@ -99,7 +97,6 @@ jobs:
               - "packages/shared/**"
             cli:
               - "packages/cli/**"
-              - "packages/runtime-mcp/**"
               - "install.sh"
             sidecar:
               - "packages/whatsapp-baileys-sidecar/**"
