+++ apps/web/Dockerfile
@@ -8,7 +8,6 @@ COPY package.json bun.lock turbo.json tsconfig.base.json biome.json ./
 COPY patches/libsignal@6.0.0.patch patches/libsignal@6.0.0.patch
 COPY apps/web/package.json apps/web/package.json
 COPY packages/cli/package.json packages/cli/package.json
-COPY packages/runtime-mcp/package.json packages/runtime-mcp/package.json
 COPY packages/shared/package.json packages/shared/package.json
 COPY packages/whatsapp-baileys-sidecar/package.json packages/whatsapp-baileys-sidecar/package.json
