+++ packages/cli/package.json
@@ -11,7 +11,6 @@
 	"files": [
 		"bin",
 		"dist",
-		"runtime-mcp",
 		"egress-addon/clawdi_egress_addon.py",
 		"skills",
 		"LICENSE",
@@ -25,17 +24,16 @@
 	},
 	"scripts": {
 		"dev": "bun run src/index.ts",
-		"build": "bun run build:mcp && rm -rf dist && bun build src/index.ts --outdir dist --target node --minify --define 'process.env.CLAWDI_DEFAULT_API_URL=\"https://cloud-api.clawdi.ai\"' --define 'process.env.CLAWDI_DEFAULT_DEPLOY_API_URL=\"https://api.clawdi.ai\"'",
-		"build:mcp": "bun build ../runtime-mcp/src/index.ts --outdir runtime-mcp --target node --minify",
+		"build": "rm -rf dist && bun build src/index.ts --outdir dist --target node --minify --define 'process.env.CLAWDI_DEFAULT_API_URL=\"https://cloud-api.clawdi.ai\"' --define 'process.env.CLAWDI_DEFAULT_DEPLOY_API_URL=\"https://api.clawdi.ai\"'",
 		"build:native": "bun run scripts/build-native.mjs",
 		"build:native-release": "bun run scripts/build-native-release.mjs",
 		"package:native-release": "bun run scripts/package-native-release.mjs",
-		"build:dev": "bun run build:mcp && rm -rf dist && bun build src/index.ts --outdir dist --target node",
+		"build:dev": "rm -rf dist && bun build src/index.ts --outdir dist --target node",
 		"check:publish-manifest": "node scripts/check-publish-manifest.mjs",
 		"check:native-release": "bun run scripts/check-native-release.mjs",
 		"typecheck": "tsc --noEmit",
 		"test": "../../scripts/test.sh cli",
-		"test:internal": "bun run build:mcp && bash scripts/test-internal.sh",
+		"test:internal": "bash scripts/test-internal.sh",
 		"test:e2e": "../../scripts/test.sh cli tests/e2e",
 		"test:native-linux-lifecycle:internal": "CLAWDI_NATIVE_BINARY=${CLAWDI_NATIVE_BINARY:-dist-native/linux-x64/clawdi} bun test tests/e2e/native-installer.e2e.test.ts tests/e2e/native-daemon.e2e.test.ts",
 		"test:watch:local": "bun test --watch",
@@ -66,7 +64,6 @@
 	"devDependencies": {
 		"@clack/prompts": "^1.7.0",
 		"@clawdi/shared": "workspace:*",
-		"@clawdi/runtime-mcp": "workspace:*",
 		"@modelcontextprotocol/sdk": "^1.30.0",
 		"@types/bun": "^1.4.0",
 		"@types/node": "^24.13.3",
