--- packages/cli/src/runtime/builtin-mcp-resource.d.ts
@@ -1,4 +1,0 @@
-declare module "*runtime-mcp/index.js" {
-	const source: string;
-	export default source;
-}
