+++ packages/cli/src/runtime/runtime-bundle-v2.test.ts
@@ -1241,7 +1241,7 @@ describe("hosted runtime bundle v2", () => {
 				expect(headers.get("accept")).toBe(HOSTED_RUNTIME_BUNDLE_V2_MEDIA_TYPE);
 				expect(headers.get("if-none-match")).toBe('"bundle-1"');
 				expect(headers.get("x-clawdi-runtime-capabilities")).toBe(
-					"agent-plugins-manifest-v1, agent-plugin-github-release-source-v1, vault-local-mcp-v1",
+					"agent-plugins-manifest-v1, agent-plugin-github-release-source-v1",
 				);
 				expect(headers.get("x-clawdi-agent-plugin-proof")).toBeNull();
 				expect(headers.get("x-clawdi-runtime-generation")).toBeNull();
