--- packages/runtime-mcp/package.json
@@ -1,33 +1,0 @@
-{
-	"name": "@clawdi/runtime-mcp",
-	"version": "0.1.0",
-	"private": true,
-	"type": "module",
-	"bin": {
-		"clawdi-mcp": "dist/index.js"
-	},
-	"files": [
-		"dist",
-		"README.md"
-	],
-	"exports": {
-		"./vault-env": "./src/vault-env.ts"
-	},
-	"engines": {
-		"node": ">=24"
-	},
-	"scripts": {
-		"build": "bun build src/index.ts --outdir dist --target node --minify",
-		"typecheck": "tsc --noEmit",
-		"test:internal": "bun test src",
-		"prepublishOnly": "bun run build"
-	},
-	"devDependencies": {
-		"@clawdi/shared": "workspace:*",
-		"@modelcontextprotocol/sdk": "^1.30.0",
-		"@types/bun": "^1.4.0",
-		"@types/node": "^24.13.3",
-		"typescript": "catalog:",
-		"zod": "catalog:"
-	}
-}
