--- packages/runtime-mcp/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ESNext",
-		"module": "ESNext",
-		"moduleResolution": "bundler",
-		"strict": true,
-		"noEmit": true,
-		"skipLibCheck": true,
-		"types": ["bun", "node"]
-	},
-	"include": ["src/**/*.ts"]
-}
