+++ packages/shared/src/api/api.generated.ts
@@ -6827,8 +6827,6 @@ export interface components {
         };
         /** HostedRuntimePlatformMcpServer */
         HostedRuntimePlatformMcpServer: {
-            /** Localvault */
-            localVault?: 1 | null;
             /**
              * Platform
              * @constant
@@ -6860,8 +6858,6 @@ export interface components {
         };
         /** HostedRuntimeRemoteMcpServer */
         HostedRuntimeRemoteMcpServer: {
-            /** Localvault */
-            localVault?: 1 | null;
             /** Url */
             url: string;
             /**
